Separeted contributor functions from server script
